Land 11 specs from the sub-recipes session plus tooling churn
App features and bug fixes (each marked done in _bmad-output/implementation-artifacts):

- Purchase auto-bump: write countDate via setHours so non-UTC hosts no longer
  store under a key the inventory page never queries (fixes the "buying does
  not bump today's count" bug).
- Per-ingredient inventory-detail popup with FIFO breakdown grouped by
  supplier, hide-on-zero, and a11y-correct dialog wiring.
- Purchases history tab with grouped receipts and invoice image attach
  (manager-only). Image stored as base64 data URL per existing convention.
- Unlock-price now recomputes the displayed cost from FIFO lots in the same
  React commit as the lock toggle (no stale manual-cost residue).
- Recipe list shows a cost range for variant-only recipes instead of \$0.00,
  with explicit empty-everywhere guard and 1-decimal formatting hygiene.
- Recipe list ingredient count now unions raw + every variation (no more
  "0 ingredients" for recipes whose ingredients live on variations).
- Per-cafe enabled-units vocabulary plus a shared <UnitPicker>; six form
  sites swapped from freeform <input> to the gated picker; legacy values
  surfaced as "(custom)" for non-destructive disable.
- Action-feed margin alerts driven by Cafe.minMarginPercent (default 20%),
  with urgent-red border on outright loss vs warning-amber on thin margin,
  itemized by losing variation in one card per recipe.
- Per-ingredient display unit on the inventory tab (within-dimension only:
  mass via g, volume via mL, count via each). Display-only — storage and
  FIFO untouched.
- Sub-recipes Phase 1: schema (yieldQuantity, yieldUnit, polymorphic
  RecipeIngredient.subRecipeId), pure expansion engine with cycle detection
  and 1-min-floor removed at iteration 1, sales deduction + cost rollup +
  margin-alert integration, FK ON DELETE Restrict to prevent silent corrupt.
- Sub-recipes Phase 2: recipe-editor UI for setting yield + adding
  composite rows via picker; "📋 {name}" composite-row rendering;
  getSubRecipeOptions action filters by yield + non-discontinued + non-self.

Test status: 628 unit tests pass (110+ new since session start). Build clean
on Next.js 16.1.6 + Prisma 7.4.2.

Schema migrations included:
- 20260504012245_add_purchase_invoice_image
- 20260504050911_add_cafe_enabled_units
- 20260504064440_add_cafe_min_margin_percent
- 20260504072216_add_ingredient_display_unit
- 20260505015617_add_recipe_sub_recipes
- 20260505022120_restrict_subrecipe_deletion

Tooling churn (pre-existing in working tree, bundled per user request):
- BMAD module reorganization under .agents/, .claude/, _bmad/
- Setup module additions, plus assorted re-pathing of BMAD agent/workflow
  files. No app-code impact.

Deferred items collected in _bmad-output/implementation-artifacts/deferred-work.md
covering perf-at-scale concerns, cycle TOCTOU race, action-layer test gaps
on sub-recipes, and a handful of UX polish items.

+++ cafe-mgmt/src/actions/supplier.actions.test.ts
@@ -22,7 +22,7 @@ const purchasePayloadSchema = z.object({
   ingredientSupplierId: z.string().min(1),
   quantity: z.number().int().min(1),
   unit: z.string().min(1).max(20),
-  totalPriceInCents: z.number().int().min(0),
+  totalPriceInCents: z.number().min(0),
 });
 
 const logOutcomeSchema = z.object({
